Ref: Created bottom_card_details, scalar and cat subfolders and moved files inside
