+++ app/src/components/market/sections/common_sections/bottom_card_details/grid_transaction_details/index.tsx
