+++ app/src/components/market/sections/common_sections/bottom_card_details/scalar/market_scale/value_boxes.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { ValueBoxItem } from '../../../../util/types'
-import { VALUE_BOXES_MARGIN } from '../common_styled'
+import { ValueBoxItem } from '../../../../../../../util/types'
+import { VALUE_BOXES_MARGIN } from '../../../../../common/common_styled'
 
 import { ValueBox } from './value_box'
