+++ app/src/components/market/sections/market_buy/scalar_market_buy.tsx
@@ -45,9 +45,9 @@ import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { CurrencySelector } from '../../common/currency_selector'
-import { GridTransactionDetails } from '../../common/grid_transaction_details'
-import { MarketScale } from '../../common/market_scale'
 import { SetAllowance } from '../../common/set_allowance'
+import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
+import { MarketScale } from '../common_sections/bottom_card_details/scalar/market_scale'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
