+++ app/src/components/market/sections/market_pooling/market_pool_liquidity.tsx
@@ -50,9 +50,9 @@ import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { CurrencySelector } from '../../common/currency_selector'
-import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { SetAllowance } from '../../common/set_allowance'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
+import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
