+++ app/src/components/market/sections/market_profile/market_status/closed.tsx
@@ -26,8 +26,8 @@ import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
 import { ModalTransactionWrapper } from '../../../../modal'
 import { MarginsButton } from '../../../common/common_styled'
-import { MarketScale } from '../../../common/market_scale'
 import { ViewCard } from '../../../common/view_card'
+import { MarketScale } from '../../common_sections/bottom_card_details/scalar/market_scale'
 import MarketResolutionMessage from '../../common_sections/messages/market_resolution_message'
 import { OutcomeTable } from '../../common_sections/tables/outcome_table'
 import { MarketTopDetailsClosed } from '../../common_sections/top_card_details/market_top_details_closed'
