+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -18,8 +18,8 @@ import {
 } from '../../../../../util/types'
 import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
-import { MarketScale } from '../../../common/market_scale'
 import { ViewCard } from '../../../common/view_card'
+import { MarketScale } from '../../common_sections/bottom_card_details/scalar/market_scale'
 import { WarningMessage } from '../../common_sections/messages/warning_message'
 import { OutcomeTable } from '../../common_sections/tables/outcome_table'
 import { MarketTopDetailsOpen } from '../../common_sections/top_card_details/market_top_details_open'
