+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -45,8 +45,8 @@ import { ButtonType } from '../../../button/button_styling_types'
 import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { ModalTransactionWrapper } from '../../../modal'
 import { GenericError } from '../../common/common_styled'
-import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
+import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
