+++ app/src/components/market/sections/market_sell/scalar_market_sell.tsx
@@ -41,10 +41,10 @@ import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { GenericError } from '../../common/common_styled'
-import { GridTransactionDetails } from '../../common/grid_transaction_details'
-import { MarketScale } from '../../common/market_scale'
 import { PositionSelectionBox } from '../../common/position_selection_box'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
+import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
+import { MarketScale } from '../common_sections/bottom_card_details/scalar/market_scale'
 import { WarningMessage } from '../common_sections/messages/warning_message'
 import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
